+++ tests/test_checker/test_noqa.py
@@ -243,7 +243,8 @@ SHOULD_BE_RAISED = types.MappingProxyType(
         'WPS476': 1,
         'WPS477': 0,  # enabled only in python 3.13+
         'WPS478': 1,
-        'WPS479': 0,  # only triggers on 3.12+
+        'WPS479': 0,
+        'WPS480': 0,  # only triggers on 3.12+
         'WPS500': 1,
         'WPS501': 1,
         'WPS502': 0,  # disabled since 1.0.0