+++ wemake_python_styleguide/presets/types/file_tokens.py
@@ -20,4 +20,5 @@ PRESET: Final = (
     primitives.WrongStringTokenVisitor,
     statements.MultilineStringVisitor,
     conditions.IfElseVisitor,
+    primitives.MultilineFormattedStringTokenVisitor,
 )