+++ CHANGELOG.md
@@ -22,7 +22,8 @@ Semantic versioning in our case means:
 ### Features
 
 - Adds `WPS478`: forbids using non strict slice operations, #1011
-- Adds `WPS469`: forbids using comments inside formatted string, #3404
+- Adds `WPS479`: forbids using multiline fstrings, #3405
+- Adds `WPS480`: forbids using comments inside formatted string, #3404
 
 ### Bugfixes